+++ src/core_modules/capture-core/components/Buttons/index.js
@@ -1,3 +1,2 @@
 // @flow
-export { Button } from './Button.component';
 export { SimpleSplitButton } from './SimpleSplitButton.component';