+++ src/core_modules/capture-core/components/D2Form/field/Components/MultiOptionSetField/MultiOptionSetSelectFieldForForm.component.js
@@ -1,0 +1,58 @@
+// @flow
+import {
+    MultiSelectField,
+    withSelectMultiTranslations,
+    withGotoInterface,
+    withHideCompatibility,
+    withDefaultShouldUpdateInterface,
+    withFocusSaver,
+    withCalculateMessages,
+    withDefaultFieldContainer,
+    withLabel,
+    withDisplayMessages,
+    withFilterProps,
+    withOptionsIconElement,
+} from '../../../../FormFields/New';
+import {
+    withRequiredFieldCalculation,
+    withDisabledFieldCalculation,
+    withRulesOptionVisibilityHandler,
+} from '../internal';
+import labelTypeClasses from '../../buildField.module.css';
+
+const getFilteredProps = (props: Object) => {
+    const { formHorizontal, fieldLabelMediaBasedClass, ...passOnProps } = props;
+    return passOnProps;
+};
+
+export const MultiOptionSetSelectFieldForForm = withGotoInterface()(
+    withHideCompatibility()(
+        withDefaultShouldUpdateInterface()(
+            withDisabledFieldCalculation()(
+                withRequiredFieldCalculation()(
+                    withFocusSaver()(
+                        withCalculateMessages()(
+                            withDefaultFieldContainer()(
+                                withLabel({
+                                    onGetUseVerticalOrientation: (props: Object) => props.formHorizontal,
+                                    onGetCustomFieldLabeClass: (props: Object) =>
+                                        `${props.fieldLabelMediaBasedClass} ${labelTypeClasses.textLabel}`,
+                                })(
+                                    withDisplayMessages()(
+                                        withOptionsIconElement()(
+                                            withRulesOptionVisibilityHandler()(
+                                                withFilterProps(getFilteredProps)(
+                                                    withSelectMultiTranslations()(MultiSelectField),
+                                                ),
+                                            ),
+                                        ),
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+            ),
+        ),
+    ),
+);