+++ src/core_modules/capture-core/components/NoticeBox/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { NoticeBox } from './NoticeBox';