+++ src/core_modules/capture-core/components/SearchBox/SearchStatus/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { SearchStatus } from './SearchStatus.container';
+