+++ src/core_modules/capture-core/components/WidgetProfile/WidgetProfile.component.js
@@ -10,10 +10,16 @@ import { FlatList } from 'capture-ui';
 import { errorCreator } from 'capture-core-utils';
 import { Widget } from '../Widget';
 import { LoadingMaskElementCenter } from '../LoadingMasks';
-import { convertValue as convertClientToView } from '../../converters/clientToView';
+import { NoticeBox } from '../NoticeBox';
 import type { Props } from './widgetProfile.types';
-import { useProgram, useTrackedEntityInstances, useClientAttributesWithSubvalues, useUserRoles } from './hooks';
-import { DataEntry, dataEntryActionTypes, TEI_MODAL_STATE, getTeiDisplayName } from './DataEntry';
+import {
+    useProgram,
+    useTrackedEntityInstances,
+    useClientAttributesWithSubvalues,
+    useUserRoles,
+    useTeiDisplayName,
+} from './hooks';
+import { DataEntry, dataEntryActionTypes, TEI_MODAL_STATE, convertClientToView } from './DataEntry';
 
 const styles = {
     header: {
@@ -67,18 +73,13 @@ const WidgetProfilePlain = ({
     const loading = programsLoading || trackedEntityInstancesLoading || userRolesLoading;
     const error = programsError || trackedEntityInstancesError || userRolesError;
     const clientAttributesWithSubvalues = useClientAttributesWithSubvalues(program, trackedEntityInstanceAttributes);
-    const teiDisplayName = getTeiDisplayName(program, storedAttributeValues, clientAttributesWithSubvalues, teiId);
+    const teiDisplayName = useTeiDisplayName(program, storedAttributeValues, clientAttributesWithSubvalues, teiId);
 
     const displayInListAttributes = useMemo(() => clientAttributesWithSubvalues
         .filter(item => item.displayInList)
-        .map(({ optionSet, attribute, key, value: clientValue, valueType }) => {
-            let value;
-            if (optionSet && optionSet.id) {
-                const selectedOption = optionSet.options.find(option => option.code === clientValue);
-                value = selectedOption && selectedOption.name;
-            } else {
-                value = convertClientToView(clientValue, valueType);
-            }
+        .map((clientAttribute) => {
+            const { attribute, key } = clientAttribute;
+            const value = convertClientToView(clientAttribute);
             return {
                 attribute, key, value, reactKey: attribute,
             };
@@ -135,20 +136,23 @@ const WidgetProfilePlain = ({
                 {renderProfile()}
             </Widget>
             {showEditModal(loading, error, isEditable, modalState) && (
-                <DataEntry
-                    onCancel={() => setTeiModalState(TEI_MODAL_STATE.CLOSE)}
-                    onDisable={() => setTeiModalState(TEI_MODAL_STATE.OPEN_DISABLE)}
-                    programAPI={program}
-                    orgUnitId={orgUnitId}
-                    clientAttributesWithSubvalues={clientAttributesWithSubvalues}
-                    userRoles={userRoles}
-                    trackedEntityInstanceId={teiId}
-                    onSaveSuccessActionType={dataEntryActionTypes.TEI_UPDATE_SUCCESS}
-                    onSaveErrorActionType={dataEntryActionTypes.TEI_UPDATE_ERROR}
-                    modalState={modalState}
-                    geometry={geometry}
-                    trackedEntityName={trackedEntityTypeName}
-                />
+                <>
+                    <DataEntry
+                        onCancel={() => setTeiModalState(TEI_MODAL_STATE.CLOSE)}
+                        onDisable={() => setTeiModalState(TEI_MODAL_STATE.OPEN_DISABLE)}
+                        programAPI={program}
+                        orgUnitId={orgUnitId}
+                        clientAttributesWithSubvalues={clientAttributesWithSubvalues}
+                        userRoles={userRoles}
+                        trackedEntityInstanceId={teiId}
+                        onSaveSuccessActionType={dataEntryActionTypes.TEI_UPDATE_SUCCESS}
+                        onSaveErrorActionType={dataEntryActionTypes.TEI_UPDATE_ERROR}
+                        modalState={modalState}
+                        geometry={geometry}
+                        trackedEntityName={trackedEntityTypeName}
+                    />
+                    <NoticeBox formId="trackedEntityProfile-edit" />
+                </>
             )}
         </div>
     );